Move CompressCommand class to com.yahoo.imapnio.async.request package.
